CLN: Move mimedata to its own file instead of in clipboard.
